Make some platform utils shared with the library
