+++ packages/diffs/src/editor/platform.ts
@@ -1,34 +1,11 @@
-let _isMacLike: boolean | undefined = undefined;
-let _isLinux: boolean | undefined = undefined;
-let _isSafari: boolean | undefined = undefined;
+import {
+  isLinux,
+  isMacLike,
+  isSafari,
+  resetPlatformDetection,
+} from '../utils/platform';
 
-/**
- * Clears the cached platform/browser detection. Detection is memoized on first
- * call, so a test process that swaps `navigator` (e.g. to exercise Linux or
- * Safari behavior) must reset it; otherwise the value cached by an earlier test
- * leaks across tests and no longer matches the active navigator.
- */
-export function resetPlatformDetectionForTests(): void {
-  _isMacLike = undefined;
-  _isLinux = undefined;
-  _isSafari = undefined;
-}
-
-export function isMacLike(): boolean {
-  return (_isMacLike ??= /macOS|MacIntel|iPhone|iPad|iPod/i.test(
-    getPlatform()
-  ));
-}
-
-export function isLinux(): boolean {
-  return (_isLinux ??= /Linux/i.test(getPlatform()));
-}
-
-export function isSafari(): boolean {
-  return (_isSafari ??=
-    ('safari' in window && 'pushNotification' in (window as any).safari) ||
-    /^((?!chrome|android).)*safari/i.test(navigator.userAgent));
-}
+export { isLinux, isMacLike, isSafari, resetPlatformDetection };
 
 export function isPrimaryModifier(
   { metaKey, ctrlKey }: MouseEvent | KeyboardEvent,
@@ -91,11 +68,4 @@ export function isMoveCursorShortcut(
   }
 
   return undefined;
-}
-
-function getPlatform(): string {
-  const navigator = globalThis.navigator as Navigator & {
-    userAgentData?: { platform?: string };
-  };
-  return navigator?.platform ?? navigator?.userAgentData?.platform ?? 'unknown';
 }
